Add findAttribute to logs module
